MarkerShape - refactoring
