+++ src/ReactLeafletMap/layers/SvgVectorLayer/Feature/shapes/Cross.js
