+++ src/ReactLeafletMap/layers/SvgVectorLayer/Feature/shapes/Pin.js
