+++ src/ReactLeafletMap/layers/SvgVectorLayer/Feature/shapes/index.js
