+++ src/ReactLeafletMap/layers/SvgVectorLayer/Feature/shapes/svg.js
